fix: изменить mark

+++ src/main/java/ru/practicum/workshop/reviewservice/dto/ReviewDto.java
@@ -13,5 +13,5 @@ public class ReviewDto {
     String content;
     LocalDateTime createdOn;
     LocalDateTime updatedOn;
-    Long mark;
+    Integer mark;
 }
